requirements.txt
click
pyyaml
parse
matplotlib
numpy
